+++ nvblox/experiments/src/tex_integration.cpp
@@ -33,7 +33,7 @@ DEFINE_string(mesh_output_path, "", "File in which to save the surface mesh.");
 DEFINE_string(texture_output_path, "",
               "File in which to save the global texture.");
 
-DEFINE_double(voxel_size, 0.0f, "Voxel resolution in meters.");
+DEFINE_double(voxel_size, 0.0, "Voxel resolution in meters.");
 DEFINE_int32(tsdf_frame_subsampling, 0,
              "By what amount to subsample the TSDF frames. A subsample of 3 "
              "means only every 3rd frame is taken.");
@@ -60,7 +60,7 @@ int main(int argc, char* argv[]) {
   options.esdf_output_path = FLAGS_esdf_output_path;
   options.mesh_output_path = FLAGS_mesh_output_path;
   options.texture_output_path = FLAGS_texture_output_path;
-  options.voxel_size = FLAGS_voxel_size;
+  options.voxel_size = static_cast<float>(FLAGS_voxel_size);
   options.tsdf_frame_subsampling = FLAGS_tsdf_frame_subsampling;
   options.color_frame_subsampling = FLAGS_color_frame_subsampling;
   options.mesh_frame_subsampling = FLAGS_mesh_frame_subsampling;
